Rename clientproc file and remove text buffer prototype
